requirements.txt
requests==2.31.0
pandas==2.1.1
python-dotenv==1.0.0
tabulate==0.9.0
matplotlib==3.8.0
websockets==11.0.3
seaborn==0.12.2 